+++ src/main/java/br/com/leonardocosta/gerenciadordetarefa/domain/dto/DepartamentoProjection.java
@@ -2,6 +2,6 @@ package br.com.leonardocosta.gerenciadordetarefa.domain.dto;
 
 public interface DepartamentoProjection {
     String getDepartamento();
-    Long getQtdPessoas();
-    Long getQtdTarefas();
+    Integer getQtdPessoas();
+    Integer getQtdTarefas();
 }
